requirements.txt
# Hyperliquid Python SDK
hyperliquid-python-sdk>=0.16.0

# Data analysis and manipulation
pandas>=2.0.0
numpy>=1.24.0

# Environment variables
python-dotenv>=1.0.0

# Additional utilities for trading
requests>=2.31.0 